Update README.md
Amazon product reviews dataset is chosen for this project to perform sentiment analysis on product reviews of  products like Kindle, Fire tablet, Echo, Fire TV, and Tap. Product reviews data is analyzed to understand customer sentiments towards it’s product performance. 
The datasets for Amazon product reviews and Amazon Stock Price is obtained from Kaggle.com.  VIX data measuring the stock market's expectation of volatility is gathered from Cboe.com. This project focuses on data collected from the years 2014 to 2017. An analysis to study the impact of product review sentiment on Amazon stock price is performed.
In addition, the next step is to forecast Amazon stock price using sentiment score, VIX (Volatility Index),and previous Amazon stock price.  Multiple linear regression model is developed to forecast the stock price by considering the previous two stock prices and the other parameters.
This study is a niche application of sentiment analysis in gauging the effects of product sentiments on stocks for a big company like Amazon. One major contribution of this work is a sentiment analysis dictionary . The sentiment scores obtained from the analysis of the product reviews is a powerful indicator of stock movements and can be used to effectively leverage the prediction of short-term trends. We believe that the reason the model is able to achieve an accuracy of 60% with this  approach is that the dictionary was specifically created for this particular sector by researching and leveraging domain expertise